Moving date formatting functionality to a separate utility module
