+++ frontend/src/utils/mod.rs
@@ -1,4 +1,5 @@
 //! Contains miscellaneous utilities for the site.
 
 pub mod background;
+pub mod date;
 pub mod open_graph;
